# Improve performance of GetFlatObjectsListWithInterface()