## Description
New implementation significantly reduces CPU and memory usage.
There are two main changes:
- Use HashSet to detect reference loops
- Don't create intermediate lists and arrays when making recursive calls

Benchmark results for a smallest entity `CommandBase`: 3 times faster, 5 times less allocated memory:
```
| Method |       Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------- |-----------:|---------:|---------:|-------:|----------:|
|    Old | 3,305.9 ns | 40.85 ns | 38.21 ns | 0.4959 |   2,089 B |
|    New |   923.9 ns |  5.68 ns |  5.03 ns | 0.1011 |     424 B |
```

Benchmark results for a large entity `ExtendedCustomerOrderAggregate` with about 13000 nested entities: **19 times faster**, 2.5 times less allocated memory:
```
| Method |        Mean |    Error |   StdDev |     Gen 0 |    Gen 1 | Allocated |
|------- |------------:|---------:|---------:|----------:|---------:|----------:|
|    Old | 1,084.94 ms | 8.185 ms | 6.390 ms | 4000.0000 |        - |     20 MB |
|    New |    56.94 ms | 1.108 ms | 1.138 ms | 1700.0000 | 300.0000 |      8 MB |
```

## References
### QA-test:
### Jira-link:
### Artifact URL: