@mvktsk 2022-05-30
Review task https://virtocommerce.atlassian.net/browse/VP-8056 has been created